Add conditional eslint execution

+++ docker_goodies/docker_entry.sh
@@ -3,14 +3,16 @@
 [[ `grep -cm1 docker /proc/1/cgroup 2> /dev/null` ]] || { echo "* Cannot execute outside Docker"; exit 1; }
 echo -e "\t+ Running coverage checks..."
 rm -rf .cov_temp
-NODE_PATH="/tmp/node_modules"
-$NODE_PATH/babel-cli/bin/babel-node.js \
-   $NODE_PATH/istanbul/lib/cli.js cover \
+NODE_MODULES_PATH="/tmp/node_modules"
+$NODE_MODULES_PATH/babel-cli/bin/babel-node.js \
+   $NODE_MODULES_PATH/istanbul/lib/cli.js cover \
    --config .istanbul-config.yml \
-   $NODE_PATH/mocha/bin/_mocha -- --recursive &> /dev/null
+   $NODE_MODULES_PATH/mocha/bin/_mocha -- --recursive &> /dev/null
 
-$NODE_PATH/istanbul/lib/cli.js check-coverage --root .cov_temp --config .istanbul-config.yml &> .cov_temp/coverage-summary
+$NODE_MODULES_PATH/istanbul/lib/cli.js check-coverage --root .cov_temp --config .istanbul-config.yml &> .cov_temp/coverage-summary
 
-echo -e "\t+ Running lint checks..."
-cat /dev/null > .lint_out
-$NODE_PATH/eslint/bin/eslint.js -f compact -o .lint_out .
+if [ -e .eslintrc ]; then
+   echo -e "\t+ Running lint checks..."
+   cat /dev/null > .lint_out
+   $NODE_MODULES_PATH/eslint/bin/eslint.js -f compact -o .lint_out .
+fi
